requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
astroid==1.6.4
async-timeout==2.0.1
atomicwrites==1.1.5
attrs==17.4.0
autobahn==18.5.1
Automat==0.6.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.3
certifi==2019.9.11
channels==1.1.8
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
coverage==4.5.1
daphne==1.4.2
decorator==4.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.6
django-allauth==0.41.0
django-appconf==1.0.3
django-autoslug==1.9.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-datetimepicker==1.5
django-bootstrap-modal-forms==1.5.0
django-bootstrap3==11.1.0
django-bootstrap3-datetimepicker-2==2.4.2
django-cities==0.5.0.4
django-ckedit==1.0
django-ckeditor==5.2.2
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-debug-toolbar==1.9.1
django-el-pagination==3.2.4
django-friendship==1.5.0
django-localflavor==1.0
django-markdown-deux==1.0.5
django-markdownx==2.0.21
django-messages==0.5.3
django-model-utils==3.0.0
django-pagedown==0.1.1
django-registration-redux==2.4
django-select2==7.1.1
django-summernote==0.8.11.4
django-taggit==0.22.2
django-test-plus==1.0.22
django-widget-tweaks==1.4.5
easy-thumbnails==2.4.1
enum34==1.1.6
ez-setup==0.9
factory-boy==2.11.1
Faker==0.8.15
flake8==3.5.0
funcsigs==1.0.2
futures==3.1.1
geoip2==2.9.0
googlemaps==3.1.3
haikunator==1.0.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
ipdb==0.11
ipython==5.7.0
ipython-genutils==0.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.8
markdown2==2.3.1
maxminddb==1.4.1
mccabe==0.6.1
more-itertools==4.2.0
msgpack-python==0.4.7
oauthlib==3.1.0
pathlib==1.0.1
pathlib2==2.3.2
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
public==1.0.3
py==1.5.3
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==1.8.2
pylint-common==0.2.5
pylint-django==0.9.0
pylint-plugin-utils==0.2.6
pypiwin32==223
pytest==3.6.0
pytest-cov==2.5.1
pytest-django==2.9.1
pytest-sugar==0.9.1
python-dateutil==2.7.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.4
pywin32==223
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.3.0
scandir==1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.4
static3==0.7.0
stripe==1.19.0
swapper==1.0.0
termcolor==1.1.0
text-unidecode==1.2
tqdm==4.11.2
traitlets==4.3.2
Twisted==19.7.0
txaio==2.10.0
Unipath==1.1
urllib3==1.25.3
virtualenv==15.2.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
win-unicode-console==0.5
wrapt==1.10.11
zope.interface==4.1.3
